Added event to reset
Added event id = ctr_events.99 to reset all holy site modifiers and
clear global flags so that the hidden events can trigger again to add
the modifiers back. This is to handle updates to holy site provinces
when moved, as the modifiers are saved in the saves, which means that if
I update the mod, the save file of those playing on previous versions of
the mod still have the modifiers on the old provinces, which now don't
match with the decisions and such.