Finishing up
Added decisions to Cathar, Waldensian & Lollard to spawn a Religious Center (Center of Reformation) in your capital.
Added a few decisions to allow some small european pirate nations to convert to old european religions.
Updated the Sikh holy site events that spawn holy sites, so that there is less of a chance of holy sites bordering eachother.
Fixed a "bug" which allowed the reformation religions to keep spawning holy sites, as long as you could keep spawning Centers of Reformations.